refactor: :art: reorganize components
