+++ client/src/components/Pages/staticPages/common/Accordion/Accordion.module.scss
