+++ client/src/components/Pages/staticPages/common/Accordion/Accordion.tsx
