+++ client/src/components/Pages/staticPages/common/Accordion/index.ts
