+++ client/src/components/Pages/staticPages/common/AutoplayVideo/AutoplayVideo.module.scss
