+++ client/src/components/Pages/staticPages/common/AutoplayVideo/index.ts
