+++ client/src/components/Pages/staticPages/common/Card/Card.tsx
