+++ client/src/components/Pages/staticPages/common/Card/index.ts
