+++ client/src/components/Pages/staticPages/common/InlineLink/InlineLink.module.scss
