+++ client/src/components/Pages/staticPages/common/InlineLink/InlineLink.tsx
