+++ client/src/components/Pages/staticPages/common/InlineLink/index.ts
