+++ client/src/components/Pages/staticPages/common/Register/Register.tsx
