+++ client/src/components/Pages/staticPages/common/Register/index.ts
