+++ client/src/components/Pages/staticPages/common/SecondaryNavbar/SecondaryNavbar.module.scss
