+++ client/src/components/Pages/staticPages/common/StepContent/StepContent.module.scss
