+++ client/src/components/Pages/staticPages/common/StepContent/index.ts
