+++ client/src/components/Pages/staticPages/common/index.ts
@@ -1,0 +1,7 @@
+export { default as Accordion } from "./Accordion";
+export { default as AutoplayVideo } from "./AutoplayVideo";
+export { default as Card } from "./Card";
+export { default as InlineLink } from "./InlineLink";
+export { default as Register } from "./Register";
+export { default as SecondaryNavbar } from "./SecondaryNavbar";
+export { default as StepContent } from "./StepContent";
