+++ client/src/components/Pages/staticPages/publier/CardExample/CardExample.tsx
