+++ client/src/components/Pages/staticPages/publier/CardExample/index.ts
