+++ client/src/components/Pages/staticPages/publier/TestimonySlider/TestimonyAuthor/index.ts
