+++ client/src/components/Pages/staticPages/publier/TestimonySlider/index.ts
