+++ client/src/components/Pages/staticPages/publier/WriteContentCard/WriteContentCard.module.scss
