+++ client/src/components/Pages/staticPages/publier/WriteContentCard/index.ts
