+++ client/src/components/Pages/staticPages/traduire/LanguageIcon/LanguageIcon.module.scss
