+++ client/src/scss/components/staticPages.module.scss
@@ -5,6 +5,7 @@
 
 $breakpoint: "lg-limit";
 
+// Layout
 .main {
   width: 100%;
 }
@@ -59,6 +60,18 @@ $breakpoint: "lg-limit";
   color: $white;
 }
 
+.anchor {
+  position: absolute;
+  top: -50px;
+}
+.top_space {
+  margin-top: u(20);
+
+  @include isTablet() {
+    margin-top: u(10);
+  }
+}
+
 // Titles
 .title2 {
   @include text("h3");
@@ -87,7 +100,82 @@ $breakpoint: "lg-limit";
   }
 }
 
-.anchor {
-  position: absolute;
-  top: -50px;
+// Hero
+.hero {
+  @include isTablet() {
+    text-align: center;
+
+    h1 {
+      @include text("h3");
+    }
+  }
+}
+.hero_title {
+  @include media-min($breakpoint) {
+    padding-right: u(12);
+  }
+  @include isTablet() {
+    margin-bottom: u(10);
+  }
+}
+
+.arrow {
+  margin-top: u(6);
+  margin-left: 80px;
+  display: flex;
+
+  @include isTablet() {
+    margin-left: inherit;
+    justify-content: center;
+  }
+
+  .arrow_btn {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 48px;
+    height: 48px;
+    background: $white;
+    border-radius: 50%;
+    animation: slide-bottom 2s cubic-bezier(0.25, 0.46, 0.45, 0.94) infinite alternate both;
+  }
+}
+
+@keyframes slide-bottom {
+  0% {
+    -webkit-transform: translateY(0);
+    transform: translateY(0);
+  }
+  100% {
+    -webkit-transform: translateY(30px);
+    transform: translateY(30px);
+  }
+}
+
+// Steps
+.warning {
+  @include text("large");
+  padding: u(4);
+  background: $attention;
+  border-radius: $radius;
+}
+.warning_mobile {
+  @extend .warning;
+  display: none;
+  align-items: center;
+  margin-bottom: u(10);
+
+  & > p {
+    margin-bottom: 0;
+    margin-left: u(4);
+  }
+
+  @include isTablet() {
+    display: flex;
+  }
+}
+
+// Faq
+.faq {
+  max-width: 720px;
 }
