+++ client/src/scss/pages/traduire.module.scss
@@ -6,89 +6,6 @@
 
 $breakpoint: "lg-limit";
 
-// Hero
-.hero {
-  @include isTablet() {
-    text-align: center;
-
-    h1 {
-      @include text("h3");
-    }
-  }
-}
-.hero_title {
-  @include media-min($breakpoint) {
-    padding-right: u(12);
-  }
-  @include isTablet() {
-    margin-bottom: u(10);
-  }
-}
-
-.arrow {
-  margin-top: u(6);
-  margin-left: 80px;
-  display: flex;
-
-  @include isTablet() {
-    margin-left: inherit;
-    justify-content: center;
-  }
-
-  .arrow_btn {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 48px;
-    height: 48px;
-    background: $white;
-    border-radius: 50%;
-    animation: slide-bottom 2s cubic-bezier(0.25, 0.46, 0.45, 0.94) infinite alternate both;
-  }
-}
-
-@keyframes slide-bottom {
-  0% {
-    -webkit-transform: translateY(0);
-    transform: translateY(0);
-  }
-  100% {
-    -webkit-transform: translateY(30px);
-    transform: translateY(30px);
-  }
-}
-
-// steps
-.warning {
-  @include text("large");
-  padding: u(4);
-  background: $attention;
-  border-radius: $radius;
-}
-.warning_mobile {
-  @extend .warning;
-  display: none;
-  align-items: center;
-  margin-bottom: u(10);
-
-  & > p {
-    margin-bottom: 0;
-    margin-left: u(4);
-  }
-
-  @include isTablet() {
-    display: flex;
-  }
-}
-// who
-.who {
-  margin-top: u(20);
-
-  @include isTablet() {
-    margin-top: u(10);
-  }
-}
-
 .next {
   h2 {
     margin-bottom: u(8);
@@ -97,41 +14,4 @@ $breakpoint: "lg-limit";
   p {
     @include text("large");
   }
-}
-
-// help
-.help {
-  margin-top: u(20);
-
-  @include isTablet() {
-    margin-top: u(10);
-  }
-}
-
-// figures
-.figures {
-  text-align: center;
-}
-.figure {
-  @include text("h1");
-  margin-bottom: u(6);
-  font-weight: bold;
-
-  @include isTablet() {
-    margin-bottom: u(4);
-  }
-}
-.figure_label {
-  @include text("h5");
-  margin-bottom: 0;
-
-  @include isTablet() {
-    @include text("large");
-    margin-bottom: u(6);
-  }
-}
-
-// faq
-.faq {
-  max-width: 720px;
 }
